made panels resizeables
