+++ package.json
@@ -16,6 +16,7 @@
     "next": "14.2.16",
     "react": "^18",
     "react-dom": "^18",
+    "react-resizable-panels": "^2.1.7",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7"
   },
